// src/shared/types/router.ts
export enum AppRoutes {
  MAIN = 'main',
  ABOUT = 'about',
  PROFILE = 'profile', // + id
  ARTICLES = 'articles',
  ARTICLE_DETAILS = 'article-page',
  ARTICLE_CREATE = 'article-create',
  ARTICLE_EDIT = 'article-edit',
  ADMIN_PANEL = 'admin-panel',

  FORBIDDEN = 'forbidden',
  NOT_FOUND = 'not_found'
}